delete dup LifeItemEntity
