+++ lib/api/firestore/life_item.dart
@@ -10,11 +10,11 @@ part 'life_item.g.dart';
 @freezed
 abstract class LifeItemEntity implements _$LifeItemEntity, Entity {
   const factory LifeItemEntity({
-    @required String type,
+    @required LifeItemType type,
     @required String key,
     @required int amount,
-    @required @TimestampConverter() DateTime createdAt,
-    @required @TimestampConverter() DateTime updatedAt,
+    @TimestampConverter() DateTime createdAt,
+    @TimestampConverter() DateTime updatedAt,
   }) = _LifeItemEntity;
   const LifeItemEntity._();
 
@@ -35,4 +35,50 @@ class LifeItemFieldEntity {
 
   /// 数量
   static const int = 'amount';
+}
+
+/// アイテム種別
+enum LifeItemType {
+  /// 職業
+  ///
+  /// 詳細は key 文字列で表現する.
+  job,
+
+  /// 株
+  stock,
+
+  /// 配偶者
+  spouse,
+
+  /// 家
+  house,
+
+  /// 金
+  money,
+
+  /// 乗り物
+  vehicle,
+
+  /// 子供(女の子)
+  childGirl,
+
+  /// 子供(男の子)
+  childBoy,
+
+  /// 火災保険
+  fireInsurance,
+
+  /// 生命保険
+  lifeInsurance,
+
+  /// 地震保険
+  earthquakeInsurance,
+
+  /// 自動車保険
+  carInsurance,
+
+  /// コーヒー
+  ///
+  /// 特殊アイテム. 所有している場合、強制的に消費して「1回休み」をくらう.
+  coffee,
 }
