+++ lib/models/play_room/life_stage.dart
@@ -2,8 +2,8 @@ import 'package:firestore_ref/firestore_ref.dart';
 import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+import '../../api/firestore/life_item.dart';
 import '../../api/firestore/user.dart';
-import '../../entities/life_item.dart';
 import '../../entities/life_step_entity.dart';
 
 part 'life_stage.freezed.dart';
