+++ test/unit/models/gain_life_item_params_test.dart
@@ -1,4 +1,4 @@
-import 'package:HumanLifeGame/entities/life_item.dart';
+import 'package:HumanLifeGame/api/firestore/life_item.dart';
 import 'package:HumanLifeGame/models/common/life_event_params/gain_life_items_params.dart';
 import 'package:HumanLifeGame/models/common/life_event_params/target_life_item_params.dart';
 import 'package:flutter_test/flutter_test.dart';
