+++ test/unit/models/life_stage_model_test.dart
@@ -1,5 +1,5 @@
+import 'package:HumanLifeGame/api/firestore/life_item.dart';
 import 'package:HumanLifeGame/api/firestore/store.dart';
-import 'package:HumanLifeGame/entities/life_item.dart';
 import 'package:HumanLifeGame/models/play_room/life_stage.dart';
 import 'package:cloud_firestore_mocks/cloud_firestore_mocks.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -9,11 +9,11 @@ import '../../helper/firestore/user_helper.dart';
 void main() {
   test('totalMoney', () async {
     const items = [
-      LifeItemEntity('doctor', LifeItemType.job, 1),
-      LifeItemEntity('money', LifeItemType.money, 200),
-      LifeItemEntity('money', LifeItemType.money, 300),
-      LifeItemEntity('money', LifeItemType.money, 100),
-      LifeItemEntity('coffee', LifeItemType.coffee, 1),
+      LifeItemEntity(key: 'doctor', type: LifeItemType.job, amount: 1),
+      LifeItemEntity(key: 'money', type: LifeItemType.money, amount: 200),
+      LifeItemEntity(key: 'money', type: LifeItemType.money, amount: 300),
+      LifeItemEntity(key: 'money', type: LifeItemType.money, amount: 100),
+      LifeItemEntity(key: 'coffee', type: LifeItemType.coffee, amount: 1),
     ];
     final firestore = MockFirestoreInstance();
     final store = Store(firestore);
