+++ test/widget/life_stage_test.dart
@@ -1,6 +1,6 @@
+import 'package:HumanLifeGame/api/firestore/life_item.dart';
 import 'package:HumanLifeGame/api/firestore/store.dart';
 import 'package:HumanLifeGame/api/firestore/user.dart';
-import 'package:HumanLifeGame/entities/life_item.dart';
 import 'package:HumanLifeGame/human_life_game_app.dart';
 import 'package:HumanLifeGame/i18n/i18n.dart';
 import 'package:HumanLifeGame/models/play_room/play_room_notifier.dart';
@@ -35,7 +35,7 @@ Future<void> main() async {
     final store = Store(firestore);
     final playRoomNotifier = PlayRoomNotifier(i18n, dice, await createPlayRoom(store));
     for (final lifeStage in playRoomNotifier.value.lifeStages) {
-      lifeStage.lifeItems.add(const LifeItemEntity('key', LifeItemType.money, 200));
+      lifeStage.lifeItems.add(const LifeItemEntity(key: 'key', type: LifeItemType.money, amount: 200));
     }
     await tester.pumpWidget(testableApp(
       home: ChangeNotifierProvider(create: (_) => playRoomNotifier, child: const LifeStages()),
